+++ comp/base/mem/mem_clear/Modules.tcl
@@ -1,0 +1,11 @@
+# Modules.tcl: Components include script
+# Copyright (C) 2024 CESNET
+# Author(s): Lukas Nevrkla <xnevrk03@stud.fit.vutbr.cz>
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+# Packages
+lappend PACKAGES "$OFM_PATH/comp/base/pkg/math_pack.vhd"
+lappend PACKAGES "$OFM_PATH/comp/base/pkg/type_pack.vhd"
+
+lappend MOD "$ENTITY_BASE/mem_clear.vhd"