+++ comp/debug/histogramer/uvm/tbench/env/pkg.sv
@@ -1,0 +1,22 @@
+//-- pkg.sv: Package for environment
+//-- Copyright (C) 2022 CESNET z. s. p. o.
+//-- Author: Lukas Nevrkla <xnevrk03@stud.fit.vutbr.cz>
+
+//-- SPDX-License-Identifier: BSD-3-Clause
+
+`ifndef UVM_HISTOGRAMER
+`define uvm_histogramer
+
+package uvm_histogramer;
+
+    `include "uvm_macros.svh"
+    import uvm_pkg::*;
+
+    `include "sequencer.sv"
+    `include "model.sv"
+    `include "scoreboard.sv"
+    `include "env.sv"
+
+endpackage
+
+`endif